make sub components for statistics
