+++ resources/js/components/statistics/TotalCard.vue
@@ -1,0 +1,30 @@
+<template>
+    <Card>
+        <template #content>
+            <div class="flex flex-wrap">
+                <span class="w-full font-bold text-xl">{{ "Total" }}</span>
+                <span class="w-20 text-muted-color-emphasis">{{ "Photos" }}:</span>
+                <span class="w-[calc(100%-5rem)] font-bold">{{ props.total.num_photos }}</span>
+                <template v-if="props.total.num_albums > 1">
+                <span class="w-20 text-muted-color-emphasis">{{ "Albums" }}:</span>
+                <span class="w-[calc(100%-5rem)] font-bold">{{ props.total.num_albums }}</span>
+                </template>
+                <span class="w-20 text-muted-color-emphasis">{{ "Size" }}:</span>
+                <span class="w-[calc(100%-5rem)] font-bold">{{ sizeToUnit(props.total.size) }}</span>
+            </div>
+        </template>
+    </Card>
+</template>
+<script setup lang="ts">
+import { sizeToUnit } from '@/utils/StatsSizeVariantToColours';
+import Card from 'primevue/card';
+
+export type TotalAlbum = {
+        num_photos: number;
+        num_albums: number;
+        size: number;
+    };
+
+const props = defineProps<{ total: TotalAlbum }>();
+
+</script>
